v02.3: Badges and More! - new room blocked with eternal flame - fire damage scales slightly with depth - new mini-bosses in infinite levels - fixed some badge balance issues - standardized graphics on some badges - new badge just for Beta Testers - donation rewards are now easier to reach - throwing potions at mobs affect them - ring of searching now affects search speed - flavor text on tombstones - Endless mode goes past depth 31 - stop the hero from ascending in Endless mode

+++ src/main/java/com/shatteredpixel/pixeldungeonunleashed/effects/Speck.java
@@ -227,7 +227,7 @@ public class Speck extends Image {
 			scale.set( Random.Float( 1, 2 ) );
 			speed.set( 0, 64 );
 			lifespan = 0.2f;
-			y -= speed.y * lifespan;
+			y -= speed.y * lifespan; // DSM-xxxx look into this one...
 			break;
 			
 		case NOTE:
@@ -394,7 +394,7 @@ public class Speck extends Image {
 				break;
 				
 			case CHANGE:
-				am = (float)FloatMath.sqrt( (p < 0.5f ? p : 1 - p) * 2);
+				am = FloatMath.sqrt( (p < 0.5f ? p : 1 - p) * 2);
 				scale.y = (1 + p) * 0.5f;
 				scale.x = scale.y * FloatMath.cos( left * 15 );
 				break;
